refactor(bot): extract bot/skill_orchestration.py — partial close #119
Sixth slice of the bot/core decomposition (PRD #113, ADR 0001). **Reduced
scope** from the original issue spec — moves only the externally-tested
public formatters; defers the internal-only orchestration helpers to a
follow-up.

Module surface (extracted in this slice):

* ``OutputMediaPaths`` dataclass + ``_collect_output_media_paths`` —
  scans an output directory for figure/table/notebook artifacts and
  classifies them by ``media_items`` type.
* ``_format_auto_route_banner`` / ``_format_auto_disambiguation`` /
  ``_AUTO_DISAMBIGUATE_GAP`` — the auto-routing UX layer that emits
  the "auto-routed to X" banner or the disambiguation block when the
  capability resolver's top-2 candidates are close.

Late import: ``_format_auto_disambiguation`` reaches into
``bot.core._skill_registry()`` for skill-description lookups, deferred
to call time.

Backward compat: ``bot.core.OutputMediaPaths``,
``bot.core._collect_output_media_paths``,
``bot.core._format_auto_route_banner``,
``bot.core._format_auto_disambiguation``,
``bot.core._AUTO_DISAMBIGUATE_GAP`` all resolve through re-export.
The 5 tests in ``tests/test_bot_output_media_paths.py``,
``tests/test_routing_regression.py``,
``tests/test_auto_routing_disambiguation.py`` continue to pass without
modification — that's the regression net.

**Deviation from issue spec — deferred to follow-up**:

The issue text also lists ``_run_omics_skill_step``,
``_run_skill_via_shared_runner``, ``_normalize_extra_args``,
``_lookup_skill_info``, ``_resolve_param_hint_info``,
``_infer_skill_for_method``, ``_build_method_preview``,
``_build_param_hint``, ``_auto_capture_dataset``,
``_auto_capture_analysis``, the env-error parsing helpers
(``_extract_env_snippet``, ``_extract_fix_hint``,
``_classify_env_error``), ``_resolve_last_output_dir``,
``_read_result_json``, ``_update_preprocessing_state``, and
``_format_next_steps``. Those are **internal helpers with no external
test imports** — moving them all in one commit would require many
late-import surgery passes for cross-module deps (registry, OUTPUT_DIR,
transcript_store, build_system_prompt, get_tool_registry, …) without
gaining any verified protection from the regression net. Once the
agent-loop slice (#121) consolidates the LLM client globals, a
follow-up issue can land that bulk move with a clearer dep boundary.

Issue #119 stays open and the follow-up scope will be tracked there.

Verification:

* Bot regression net (10 test files + ``tests/bot/`` +
  ``tests/runtime/preflight/``): 225 passed, 1 skipped, 4 xfailed —
  no regressions.
* ``tests/eval/`` + ``tests/test_knowhow.py``: 121 passed, 17
  deselected. Eval baseline 14/15 against ``deepseek-v4-pro``
  unaffected.
* ``bot/core.py``: 3880 → 3811 LOC (−69 in this slice).
* ``bot/skill_orchestration.py`` lands at 116 LOC.
* Cumulative drop from main: 5160 → 3811 (−1349, −26.1%).

Refs #119.
